Factored out WebChat hooks into separate headless component
