+++ Composer/packages/client/src/components/WebChat/hooks/WebChatHooksContainer.tsx
@@ -1,0 +1,14 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import React from 'react';
+
+import { useActivityInspectionListener } from './useActivityInspectionListener';
+import { useTranscriptFocusListener } from './useTranscriptFocusListener';
+
+export const WebChatHooksContainer: React.FC<{}> = () => {
+  useActivityInspectionListener();
+  useTranscriptFocusListener();
+
+  return null;
+};
